move fixed_string from compile.h to format.h
